pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.4",
    "pyside6>=6.7.0",
    "rapidfuzz>=3.9.0",
    "pillow>=10.3.0",
    "pytesseract>=0.3.10",
    "mss>=9.0.1",
    "keyboard>=0.13.5",
    "pynput>=1.7.7",
    "numpy>=2.1.1",
    "rapidocr-onnxruntime>=1.3.19",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.11.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.scripts]
build-exe = "python build.py"

[tool.setuptools.packages.find]
where = ["."]
include = ["atlas", "atlas.*", "maps"]
exclude = ["static*", "debug*", "dist*", "docs*", "tools*", "__pycache__*"]
